Move getters and setters to trait
