+++ src/Message/PurchaseRequest.php
@@ -3,10 +3,13 @@
 namespace Omnipay\PaymentVision\Message;
 
 use Omnipay\Common\Message\AbstractRequest;
+use Omnipay\PaymentVision\CommonParametersTrait;
 use Omnipay\PaymentVision\CreditCardHelper;
 
 class PurchaseRequest extends AbstractRequest
 {
+    use CommonParametersTrait;
+
     /**
      * SoapClient Class
      */
@@ -15,121 +18,6 @@ class PurchaseRequest extends AbstractRequest
     public function getSoap()
     {
         return $this->soap;
-    }
-
-    public function getPvLogin()
-    {
-        return $this->getParameter('pvLogin');
-    }
-
-    public function setPvLogin($value)
-    {
-        return $this->setParameter('pvLogin', $value);
-    }
-
-    public function getPvPassword()
-    {
-        return $this->getParameter('pvPassword');
-    }
-
-    public function setPvPassword($value)
-    {
-        return $this->setParameter('pvPassword', $value);
-    }
-
-    public function getPvAPIKey()
-    {
-        return $this->getParameter('pvAPIKey');
-    }
-
-    public function setPvAPIKey($value)
-    {
-        return $this->setParameter('pvAPIKey', $value);
-    }
-
-    public function getPvToken()
-    {
-        return $this->getParameter('pvToken');
-    }
-
-    public function setPvToken($value)
-    {
-        return $this->setParameter('pvToken', $value);
-    }
-
-    public function getMerchantPayeeCode()
-    {
-        return $this->getParameter('merchantPayeeCode');
-    }
-
-    public function setMerchantPayeeCode($value)
-    {
-        return $this->setParameter('merchantPayeeCode', $value);
-    }
-
-    public function getComment()
-    {
-        return $this->getParameter('comment');
-    }
-
-    public function setComment($value)
-    {
-        return $this->setParameter('comment', $value);
-    }
-
-    public function getUserDefinedOne()
-    {
-        return $this->getParameter('userDefinedOne');
-    }
-
-    public function setUserDefinedOne($value)
-    {
-        return $this->setParameter('userDefinedOne', $value);
-    }
-
-    public function getHoldForApproval()
-    {
-        return $this->getParameter('holdForApproval');
-    }
-
-    public function setHoldForApproval($value)
-    {
-        return $this->setParameter('holdForApproval', $value);
-    }
-
-    public function getIsRecurring()
-    {
-        return $this->getParameter('isRecurring');
-    }
-
-    public function setIsRecurring($value)
-    {
-        return $this->setParameter('isRecurring', $value);
-    }
-
-    public function getLiveWsdl()
-    {
-        return $this->getParameter('liveWsdl');
-    }
-
-    public function setLiveWsdl($value)
-    {
-        return $this->setParameter('liveWsdl', $value);
-    }
-
-    public function getTestWsdl()
-    {
-        return $this->getParameter('testWsdl');
-    }
-
-    public function setTestWsdl($value)
-    {
-        return $this->setParameter('testWsdl', $value);
-    }
-
-    public function getWsdl()
-    {
-        return $this->getTestMode() ? $this->getTestWsdl() : $this->getLiveWsdl();
     }
 
     /**
